requirements.txt
matplotlib
numpy
pandas
scipy 
scikit-learn
PyYAML
-e .
six